в43/requirements.txt
# Основные зависимости
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
asyncio>=3.4.3
python-dotenv>=1.0.0
flask>=2.3.0
flask-login>=0.6.2
flask-socketio>=5.3.0
python-socketio>=5.9.0

# ML/AI
scikit-learn>=1.3.0
tensorflow>=2.13.0
xgboost>=1.7.0
optuna>=3.3.0
ta-lib>=0.4.28

# NLP и анализ текста  
transformers>=4.30.0
nltk>=3.8.0
textblob>=0.17.0
torch>=2.0.0  # Для transformers

# Социальные сети
tweepy>=4.14.0
praw>=7.7.0

# Визуализация
plotly>=5.15.0
dash>=2.11.0

# Парсинг новостей
feedparser>=6.0.10
newspaper3k>=0.2.8
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Утилиты
python-dateutil>=2.8.2
pytz>=2023.3
aiohttp>=3.8.5
redis>=4.6.0  # Для кеширования
APScheduler>=3.10.0  # Для планировщика задач